+++ oid4vc-core/src/utils/mod.rs
@@ -1,1 +1,1 @@
-pub mod to_form_urlencoded_string;
+pub mod form_urlencoded;
