+++ oid4vc-manager/tests/oid4vp/mod.rs
@@ -1,2 +1,1 @@
 pub mod implicit;
-pub mod old_implicit;
